+++ src/Controller/ManufacturersController.php
@@ -98,7 +98,7 @@ class ManufacturersController extends AppController
          //       $this->Flash->error(__('The manufacturer could not be saved. Please, try again.'));
             }
         }
-        $customers = $this->Manufacturers->Customers->find('list', ['limit' => 200]);
+        $customers = $this->Manufacturers->Customers->find('list', ['limit' => 200])->where(['customer_id' => $this->loggedinuser['customer_id']])->orwhere(['customer_id' => '0']) ;
         $this->set(compact('manufacturer', 'customers'));
         $this->set('_serialize', ['manufacturer']);
     }
@@ -115,6 +115,12 @@ class ManufacturersController extends AppController
         $manufacturer = $this->Manufacturers->get($id, [
             'contain' => []
         ]);
+		if($manufacturer['customer_id']!= $this->loggedinuser['customer_id'])
+		{
+			 $this->Flash->success(__('You are not Authorized.'));
+			 return $this->redirect(['action' => 'index']);
+			
+		}
         if ($this->request->is(['patch', 'post', 'put'])) {
             $manufacturer = $this->Manufacturers->patchEntity($manufacturer, $this->request->data);
 			$manufacturer['customer_id']=$this->loggedinuser['customer_id'];
@@ -126,7 +132,7 @@ class ManufacturersController extends AppController
          //       $this->Flash->error(__('The manufacturer could not be saved. Please, try again.'));
             }
         }
-        $customers = $this->Manufacturers->Customers->find('list', ['limit' => 200]);
+        $customers = $this->Manufacturers->Customers->find('list', ['limit' => 200])->where(['customer_id' => $this->loggedinuser['customer_id']])->orwhere(['customer_id' => '0']) ;
         $this->set(compact('manufacturer', 'customers'));
         $this->set('_serialize', ['manufacturer']);
     }
@@ -142,12 +148,19 @@ class ManufacturersController extends AppController
     {
        // $this->request->allowMethod(['post', 'delete']);
         $manufacturer = $this->Manufacturers->get($id);
-        if ($this->Manufacturers->delete($manufacturer)) {
-       //     $this->Flash->success(__('The manufacturer has been deleted.'));
-        } else {
-       //     $this->Flash->error(__('The manufacturer could not be deleted. Please, try again.'));
-        }
-
+		if($manufacturer['customer_id'] = $this->loggedinuser['customer_id'])
+	    {
+		        if ($this->Manufacturers->delete($manufacturer)) {
+		       //     $this->Flash->success(__('The manufacturer has been deleted.'));
+		        } else {
+		       //     $this->Flash->error(__('The manufacturer could not be deleted. Please, try again.'));
+		        }
+		}
+		 else
+		 {
+	   	    $this->Flash->error(__('You are not authorized'));
+		
+	     }
         return $this->redirect(['action' => 'index']);
     }
