+++ src/Controller/ServicetasksController.php
@@ -99,7 +99,7 @@ class ServicetasksController extends AppController
             //    $this->Flash->error(__('The servicetask could not be saved. Please, try again.'));
             }
         }
-        $customers = $this->Servicetasks->Customers->find('list', ['limit' => 200]);
+        $customers = $this->Servicetasks->Customers->find('list', ['limit' => 200])->where(['customer_id' => $this->loggedinuser['customer_id']])->orwhere(['customer_id' => '0']) ;
         $this->set(compact('servicetask', 'customers'));
         $this->set('_serialize', ['servicetask']);
     }
@@ -116,6 +116,12 @@ class ServicetasksController extends AppController
         $servicetask = $this->Servicetasks->get($id, [
             'contain' => []
         ]);
+		if($servicetask['customer_id']!= $this->loggedinuser['customer_id'])
+		{
+			 $this->Flash->success(__('You are not Authorized.'));
+			 return $this->redirect(['action' => 'index']);
+			
+		}
         if ($this->request->is(['patch', 'post', 'put'])) {
             $servicetask = $this->Servicetasks->patchEntity($servicetask, $this->request->data);
 			$servicetask['customer_id']=$this->loggedinuser['customer_id'];
@@ -127,7 +133,7 @@ class ServicetasksController extends AppController
              //   $this->Flash->error(__('The servicetask could not be saved. Please, try again.'));
             }
         }
-        $customers = $this->Servicetasks->Customers->find('list', ['limit' => 200]);
+        $customers = $this->Servicetasks->Customers->find('list', ['limit' => 200])->where(['customer_id' => $this->loggedinuser['customer_id']])->orwhere(['customer_id' => '0']) ;
         $this->set(compact('servicetask', 'customers'));
         $this->set('_serialize', ['servicetask']);
     }
@@ -143,12 +149,19 @@ class ServicetasksController extends AppController
     {
         //$this->request->allowMethod(['post', 'delete']);
         $servicetask = $this->Servicetasks->get($id);
-        if ($this->Servicetasks->delete($servicetask)) {
-        //    $this->Flash->success(__('The servicetask has been deleted.'));
-        } else {
-        //    $this->Flash->error(__('The servicetask could not be deleted. Please, try again.'));
-        }
-
+		if($servicetask['customer_id'] = $this->loggedinuser['customer_id'])
+	    {
+	        if ($this->Servicetasks->delete($servicetask)) {
+	        //    $this->Flash->success(__('The servicetask has been deleted.'));
+	        } else {
+	        //    $this->Flash->error(__('The servicetask could not be deleted. Please, try again.'));
+	        }
+		 }
+		 else
+		 {
+	   	    $this->Flash->error(__('You are not authorized'));
+		
+	     }
         return $this->redirect(['action' => 'index']);
     }
