+++ src/Controller/SymbolsController.php
@@ -101,7 +101,7 @@ class SymbolsController extends AppController
              //   $this->Flash->error(__('The symbol could not be saved. Please, try again.'));
             }
         }
-        $customers = $this->Symbols->Customers->find('list', ['limit' => 200]);
+        $customers = $this->Symbols->Customers->find('list', ['limit' => 200])->where(['customer_id' => $this->loggedinuser['customer_id']])->orwhere(['customer_id' => '0']) ;
         $this->set(compact('symbol', 'customers'));
         $this->set('_serialize', ['symbol']);
     }
@@ -118,6 +118,12 @@ class SymbolsController extends AppController
         $symbol = $this->Symbols->get($id, [
             'contain' => []
         ]);
+		if($symbol['customer_id']!= $this->loggedinuser['customer_id'])
+		{
+			 $this->Flash->success(__('You are not Authorized.'));
+			 return $this->redirect(['action' => 'index']);
+			
+		}
         if ($this->request->is(['patch', 'post', 'put'])) {
             $symbol = $this->Symbols->patchEntity($symbol, $this->request->data);
 			$symbol['customer_id']=$this->loggedinuser['customer_id'];
@@ -129,7 +135,7 @@ class SymbolsController extends AppController
              //   $this->Flash->error(__('The symbol could not be saved. Please, try again.'));
             }
         }
-        $customers = $this->Symbols->Customers->find('list', ['limit' => 200]);
+        $customers = $this->Symbols->Customers->find('list', ['limit' => 200])->where(['customer_id' => $this->loggedinuser['customer_id']])->orwhere(['customer_id' => '0']) ;
         $this->set(compact('symbol', 'customers'));
         $this->set('_serialize', ['symbol']);
     }
@@ -145,12 +151,19 @@ class SymbolsController extends AppController
     {
         //$this->request->allowMethod(['post', 'delete']);
         $symbol = $this->Symbols->get($id);
-        if ($this->Symbols->delete($symbol)) {
-         //   $this->Flash->success(__('The symbol has been deleted.'));
-        } else {
-        //    $this->Flash->error(__('The symbol could not be deleted. Please, try again.'));
-        }
-
+		if($symbol['customer_id'] = $this->loggedinuser['customer_id'])
+	    {
+		        if ($this->Symbols->delete($symbol)) {
+		         //   $this->Flash->success(__('The symbol has been deleted.'));
+		        } else {
+		        //    $this->Flash->error(__('The symbol could not be deleted. Please, try again.'));
+		        }
+		 }
+		 else
+		 {
+	   	    $this->Flash->error(__('You are not authorized'));
+		
+	     }
         return $this->redirect(['action' => 'index']);
     }
